+++ packages/grafana-ui/src/components/TextArea/TextArea.story.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TextArea } from './TextArea';
-import { withCenteredStory } from '../../../utils/storybook/withCenteredStory';
+import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
 import { boolean, number, select, text } from '@storybook/addon-knobs';
 import mdx from './TextArea.mdx';
