+++ packages/grafana-ui/src/components/index.ts
@@ -138,6 +138,7 @@ export { RadioButtonGroup } from './Forms/RadioButtonGroup/RadioButtonGroup';
 
 export { Input } from './Input/Input';
 export { Switch } from './Forms/Switch';
+export { TextArea } from './TextArea/TextArea';
 
 // Legacy forms
